# Add light weight union ptr equality check 