It's not unusual when building compilers to do transformations which require collecting sets over an AST – but you'll often need to do branching and unions across very similar sets, for example, one side added 2 items into a set of 1000. This is currently quite slow.

Adding a light weight pointer equality check can actually help quite a lot in these cases.

If I add some small additional benchmarks to union as such:
```haskell
defaultMain
      [ bench "union_same" $ whnf (S.union s_even) s_even
      , bench "union_minor_diff" $ whnf (S.union s_even) (S.insert 1 s_even)
      , bench "union" $ whnf (S.union s_even) s_odd
      ]
```


I'll see these results:

```
Before:
  union_same:       OK
    11.4 μs ± 863 ns
  union_minor_diff: OK
    11.3 μs ± 850 ns
  union:            OK
    77.9 μs ± 7.0 μs

After:
  union_same:       OK
    2.02 ns ± 138 ps
  union_minor_diff: OK
    137  ns ± 7.2 ns
  union:            OK
    77.4 μs ± 6.9 μs
```