@HuwCampbell 2025-08-13
Forgive me, there is one case where the other check (which the second commit removes) does provide an improvement.

If we imagine creating the same set in completely separate manners (i.e., they don't share any pointers) then the insertR rule will kick in and you'll get a 25% speed up saving on rebalance costs.

```
Before
  union_same_but_different:   OK
    29.0 μs ± 2.0 μs

After
  union_same_but_different:   OK
    40.0 μs ± 3.4 μs
```

I think this best case 25% improvement is massively dwarfed by the 5000% we see elsewhere.