@HuwCampbell 2025-08-14
There's a pretty reasonable question to be asked about whether this optimisation should also be applied to maps. The `insert` operation there does quite a bit dance on pointer equalities effectively just for maps structurally including the same items (and noting this is often only true for union).