@meooow25 2025-08-14
I'm not convinced that this is a good idea. You're proposing a *time optimization* that applies when the two maps being union-ed happen to have identical (by pointer) subtrees. This might be happening commonly for your case, but there is no reason to expect this in general and the price to pay is a linear overhead for everybody else.

This is also completely orthogonal to the pointer checks you have removed. Those are a *memory optimization*, to avoid allocating a result `Bin` when it would be equal to the input `Bin`.