@HuwCampbell 2025-08-15
I don't particularly mind putting back the second pointer equality checks. I just thought that with the proposed change they're less likely to be hit.

But let's talk it through. The existing code and checks are actually very good at dealing with trees of different balance, so two sets which are `Eq` equal should turn out with the balance (tree structure) of the first argument, and indeed, in the end, share a pointer (I don't have a proof of this).

So what we had before will lower allocations and be faster for union `Eq` equal sets, or sets with significant chunks which are `(==)`.

```haskell
-- These two have the same elements, but are structured differently
s1 = Set.fromList [1,4,3,2,5]
-- 3
-- +--1
-- |  +--|
-- |  +--2
-- +--4
--    +--|
--    +--5
s2 = Set.fromList [1,5,2,3,4]
-- 2
-- +--1
-- +--4
--    +--3
--    +--5
s3 = Set.union s1 s2
-- ptrEq s1 s3 should be True
```

But where do these come from? In what scenarios do we create two sets with the same elements and then union them? In my case, it was because I was working with two sets which share a common history and have been rejoined, and I hypothesised that for large sets it's probably the main way this will happen.

So my thought was that we should be able to see if this is actually test if this is the case and just short-circuit where we see it.

For performance in cases with no sharing, the cost is very modest; 77.9 to 78.1 microseconds – lets go 1%. The cost of the 2 pointer checks also checks in at about 1% each above (so 2%).

But, there are occasional _significant_ wins with 100x seen above when we can effectively short circuit out all but one path to a new element.