@meooow25 2025-08-31
Sorry, I've been a little too busy to look at this properly until now.

My main concern was with the overhead of the added check, but you're right that that does seem to be minimal.

I'm still a little uneasy about relying on pointer-equality checks, especially when it drastically affects the running time. For example, it wouldn't be great if a user makes a small change and their program becomes much slower because the check doesn't trigger now. But I haven't been able to think of such a situation so far.

This change would also open the door to similar checks in other operations (intersection, difference, isSubsetOf, etc.) and other structures, so we have to consider that too.

I'm also not aware of any other Haskell library using pointer-equality for short-circuits. That could have made the decision easier.

@treeowl do you have any opinions on this?